backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
markdown
pygments
